Add optimization (#48)
Also restructure the files.
